#62 fix/Update refont back ressources cours
