+++ server/firebase.js
@@ -3,9 +3,6 @@ const { getFirestore, FieldValue } = require("firebase-admin/firestore");
 const { getAuth } = require("firebase-admin/auth");
 const admin = require("firebase-admin");
 
-const admin = require("firebase-admin");
-
-
 const serviceAccount = require("./credentials.json");
 
 const app = initializeApp({
